# public/openapi.yaml
openapi: 3.0.3
info:
  title: Authy Platform API
  version: 1.0.0
  description: Tenant-isolated identity, application access, and audit API.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /vault/{id}:
    patch:
      summary: Update an encrypted Vault item and assignments (admin)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: Vault item updated } }
    delete:
      summary: Delete a Vault item (admin)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '204': { description: Vault item deleted } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: authy.session_token }
    bearerAuth: { type: http, scheme: bearer }
...